backend/requirements.txt
flask
pymongo
bcrypt
PyJWT
python-dotenv